+++ .github/CODEOWNERS
@@ -7,3 +7,4 @@
 /images/arch       @Foxboron
 /images/rhel       @debarshiray @olivergs
 /images/ubuntu     @Jmennius
+/src/pkg/utils/arch.go                    @Foxboron
