+++ src/meson.build
@@ -31,6 +31,7 @@ sources = files(
   'pkg/term/term.go',
   'pkg/term/term_test.go',
   'pkg/utils/libsubid-wrappers.c',
+  'pkg/utils/arch.go',
   'pkg/utils/errors.go',
   'pkg/utils/utils.go',
   'pkg/utils/utils_cgo.go',
